requirements.txt
# Core Dependencies
faster-whisper==1.0.3
pyaudio==0.2.14
pyttsx3==2.90

# Porcupine Wake Word ✅ ADDED
pvporcupine==3.0.2

# NLP & ML
scikit-learn==1.3.2
joblib==1.3.2
numpy==1.24.3

# Vision & Screen
google-generativeai==0.3.2
pillow==10.1.0
pyautogui==0.9.54
mss==9.0.1

# OmniParser Dependencies
torch==2.1.0
torchvision==0.16.0
ultralytics==8.0.196
paddlepaddle==2.5.1
paddleocr==2.7.0

# System Control
pycaw==20240210
wmi==1.5.1
comtypes==1.4.1

# Utilities
python-dotenv==1.0.0
colorama==0.4.6
